Basics

MCPhyloTree.add_child! — Method

add_child!(mother_node::N, child::N, child_position::Union{Int64, Missing}=missing) where N <: GeneralNode

This function adds a child to the mother node. The arity of the mother node is increased by 1 and the root status of the child is set to False.

mother_node : Node to add a child to.

child : Node to add to mother_node.children.

child_position : index at which to add the new child node; optional.

MCPhyloTree.check_binary — Method

check_binary(root::GeneralNode)::Bool

checks to see if given tree is binary; returns true if properly formatted and false otherwise

MCPhyloTree.check_leafsets — Method

check_leafsets(trees::Vector{T})::Nothing where T<:GeneralNode

Checks if an array of trees shares the same leafset (based on the leaf names)

MCPhyloTree.delete_node! — Method

delete_node!(node::T)::Nothing where T<:GeneralNode

This functions deletes node from a tree and assigns all its children to its mother node.

node : Node to be deleted.

MCPhyloTree.force_ultrametric! — Method

force_ultrametric!(root::T) where T<:GeneralNode

Force an ultrametric version of the tree.

MCPhyloTree.get_branchlength_vector — Method

get_branchlength_vector(root::N)::Vector{T}  where {N <:GeneralNode, T<:Real}

Get the vector of branch lengths of the tree.

MCPhyloTree.get_branchlength_vector — Method

get_branchlength_vector(root::N, out_vec::Vector{T}) where {N<:GeneralNode, T<:Real}

Do post order traversal to retrieve a vector of branch lengths.

MCPhyloTree.get_mother — Method

get_mother(node::T)::T  where T<:GeneralNode

This function gets the mother of node. It does so by looking for the respective binary representation of the mother node.

MCPhyloTree.get_sister — Method

get_sister(node::T)::T  where T<:GeneralNode

This function gets the sister of node. It does so by looking for the respective binary representation of the sister.

MCPhyloTree.get_sum_seperate_length! — Method

get_sum_seperate_length!(root::T)::Vector{Float64}  where T<:GeneralNode

This function gets the sum of the branch lengths of the internal branches and the branches leading to the leave nodes.

MCPhyloTree.get_sum_seperate_length! — Method

get_sum_seperate_length!(post_order::Vector{T})::Vector{Float64}  where T<:GeneralNode

This function gets the sum of the branch lengths of the internal branches and the branches leading to the leave nodes.

MCPhyloTree.insert_node! — Method

insert_node!(mother::T, children::Vector{T})::T where T<:GeneralNode

This function inserts a node into a tree after a mother node and gains a subset of the mother's children as its children.

Returns the inserted node.

mother : Node under which to add the newly-inserted node.

children : Children of node referenced by "mother" to reassign as children of the newly-inserted node.

MCPhyloTree.node_depth — Method

node_depth(node::T)::Int64 where T<:GeneralNode

Calculate the depth of a node.

MCPhyloTree.node_height — Method

node_height(root::T, mv::Float64)::Float64  where T<:GeneralNode

Calculate the height of a node.

MCPhyloTree.number_nodes! — Method

number_nodes!(root::T)::Nothing  where T<:GeneralNode

This function assigns a unique, sequential number to each node. Leaves are numbered first in alphabetical order.

MCPhyloTree.path_length — Method

path_length(ancestor::T, descendant::T)::Float64  where T<:GeneralNode

Note: The function assumes there is an ancestral relationship between the two nodes.

This function calculates the length of the path separating the ancestor from the offspring node. The function follows the path specified through the binary description of the node.

MCPhyloTree.random_node — Method

random_node(root::T)::T  where T<:GeneralNode

This function returns a random node from the tree.

MCPhyloTree.remove_child! — Method

remove_child!(mother_node::N, left::Bool)::N where N <: GeneralNode

This function removes a child from the list of nodes which are daughters of this node. An input of "True" removes the left child, while "False" removes the right child.

Returns the removed node.

mother_node : Node from which to remove a child.

left : boolean value determining which child of mother_node to remove.

MCPhyloTree.remove_child! — Method

remove_child!(mother_node::N, child::N)::N where N<:GeneralNode

This function removes a child from the list of nodes which are daughters of this node.

The removed node is returned.

mother_node : Node from which to remove a child.

child : specific Node to remove. This node has to be a child of mother_node.

MCPhyloTree.set_binary! — Method

set_binary!(root::T)  where T <: GeneralNode

Assign a binary representation to each node, which specifies the path from the root to this node via the binary representation of the node. A left turn is a 1 in binary and a right turn a 0.

MCPhyloTree.set_branchlength_vector! — Method

set_branchlength_vector!(root::N, blenvec::Array{T}) where {N<:GeneralNode, T<:Real}

This function sets the branch lengths of a tree to the values specified in blenvec.

MCPhyloTree.tree_height — Method

tree_height(root::T)::Float64  where T<:GeneralNode

This function calculates the tree height.

MCPhyloTree.tree_length — Method

tree_length(root::T)::Float64  where T<:GeneralNode

This function calculates the tree length.

MCPhyloTree.tree_length — Method

tree_length(root::T, tl::Float64)::Float64 where T<:GeneralNode

This function does the internal tree length recursion

Tree Pruning

MCPhyloTree.prune_tree! — Method

prune_tree!(root::T, node_names::Vector{String})::Nothing where T<:GeneralNode

In-place version of prune_tree.

root : root Node of tree to prune.

node_names : vector of strings, used to specify nodes to remove.

MCPhyloTree.prune_tree! — Method

prune_tree!(root::T, node_names::Vector{T})::Nothing where T<:GeneralNode

In-place version of prune_tree.

root : root node of tree to prune.

node_names: vector of Node objects to be removed from tree.

MCPhyloTree.prune_tree — Method

prune_tree(root::T, node_names::Vector{String})::T where T<:GeneralNode

This function removes specific nodes, including their descendants, from a tree.

root : root Node of tree to prune.

node_names : vector of strings, used to specify nodes to remove.

Tree Search

MCPhyloTree.find_binary — Method

find_binary(root::T, bin::String)::T where T<:GeneralNode

Find a node by its binary representation. The function assumes that the node is present in the tree.

Do not use this function if you are unsure whether the node is in the tree at all.

Returns a reference to the desired Node.

root : root Node of tree to search.

bin : binary representation of desired Node as a String.

MCPhyloTree.find_num — Method

find_num(root::T, num::Int64, rn::Vector{T})::Bool  where T<:GeneralNode

Do a post order traversal to find the node corresponding to the num.

Returns true if node is found, false otherwise. Desired Node is pushed to rn.

root : root Node of tree to be searched.

num : number of desired Node.

rn : Vector of Nodes; desired Node is pushed to this vector when found.

MCPhyloTree.find_num — Method

find_num(root::T, num::Int64)  where T<:GeneralNode

Find a node by its number. The function assumes that the node is present in the tree.

Do not use this function if you are unsure whether the node is in the tree at all.

Returns reference to Node.

root : root Node of tree to be searched.

num : number of desired Node.

MCPhyloTree.find_root — Method

find_root(node::T)::T where T <: GeneralNode

Finds the root of tree indicated by Node.

Returns reference to root Node of the tree.

node : Node in Tree of interest.

Tree Traversal

MCPhyloTree.get_leaves — Method

get_leaves(root::T)::Vector{T} where T<:GeneralNode

This function returns the leaves of a tree. Only the root node needs to be supplied.

Returns vector of leaf Nodes.

root : root Node of tree.

MCPhyloTree.get_leaves — Method

get_leaves(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function returns leaf nodes of a tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used.

Returns a vector of leaf Nodes.

root : root Node of tree to traverse.

traversal : vector of Nodes; leaf Nodes, once found, are pushed to this vector.

MCPhyloTree.level_order — Method

level_order(node::T)::Array{T} where T<:GeneralNode

This function does level order traversal. Only the root node needs to be supplied.

Returns Array of Nodes.

node : root Node of tree.

MCPhyloTree.level_traverse — Method

level_traverse(node::T, level::Int64, stack::Array{T})::Bool where T <:GeneralNode

This function traverses a level of the tree specified through node. The level is specified via the level argument and the nodes visited are stored in the stack. This function is intended as the internal worker for the level_order function.

MCPhyloTree.post_order — Method

post_order(root::T)::Vector{T} where T<:GeneralNode

This function does post order traversal. Only the root node needs to be supplied.

Returns vector of Nodes.

root : root Node of tree.

MCPhyloTree.post_order — Method

post_order(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function performs a post order traversal through the tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used for the post order traversal.

Returns vector of Nodes.

root : root Node of tree.

traversal : vector of Nodes; nodes are pushed to this vector as the tree is traversed.

MCPhyloTree.pre_order — Method

pre_order(root::T)::Vector{T} where T<:GeneralNode

This function does pre order traversal. Only the root node needs to be supplied.

Returns vector of Nodes.

root : root Node of tree.

MCPhyloTree.pre_order — Method

pre_order(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function performs a pre order traversal through the tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used for the pre order traversal.

Returns vector of Nodes.

root : root Node of tree.

traversal : vector of Nodes; nodes are pushed to this vector as the tree is traversed.