Building

Converter

MCPhyloTree.from_df — Method

from_df(df::Array{Float64,2}, name_list::Vector{String})::GeneralNode

This function takes an adjacency matrix and a vector of names and turns it into a tree. No checks are performed.

Returns the root node of the tree.

df : matrix with edge weights

name_list : a list of names such that they match the column indices of the matrix

MCPhyloTree.newick — Method

newick(root::T)::String  where T<:GeneralNode

Creates a newick representation of the tree.

Returns a properly formatted newick String.

node : root node of tree used to create the newick string.

MCPhyloTree.newick — Method

newick(root::T, newickstring::AbstractString) where T<:GeneralNode

Do the newick recursion. It is meant as the internal iterator function.

MCPhyloTree.to_covariance — Method

to_covariance(tree::N, blv::Array{T})::Array{T,2} where {N<:GeneralNode,T<: Real}

Calcualte the variance-covariance matrix from tree. An entry (i,j) of the matrix is defined as the length of the path connecting the latest common ancestor of i and j with the root of the tree.

Returns an Array of Real numbers.

tree : Node in tree of interest.

blv : branchlength vector of tree.

MCPhyloTree.to_df — Method

to_df(root::GeneralNode)::Tuple{Array{Float64}, Vector{String}}

This function returns a matrix representation of the tree structure and a vector with the column names. The entry mat[i,j] is the length of the edge connecting node i with node j. Returns Tuple containing the matrix and a vector of names.

root : root of tree used to create matrix represenation.

MCPhyloTree.to_distance_matrix — Method

to_distance_matrix(tree::T)::Array{Float64,2} where T <:GeneralNode

Calculate the distance matrix over the set of leaves.

Returns an Array of Floats.

tree : root node of tree used to perform caclulcation.

Newick Parsing

MCPhyloTree.ParseNewick — Method

ParseNewick(filename::String)::Array{GeneralNode, 1}

This function takes a filename as a String, and returns an array of trees(represented as Node objects). The file should solely consist of newick tree representations, separated by line. The function checks for proper newick formatting, and will return an error if the file is incorrectly formatted.

Returns an Array of Nodes; each Node is the root of the tree represented by a newick string in the file.

filename : name of file containing newick strings to parse.

Tree Building

Tree Clustering

MCPhyloTree.neighbor_joining — Method

neighbor_joining(dm::Array{Float64,2}, Array{String,1})

This function returns a phylogenetic tree by using neighbor-joining based on a given distance matrix and an array of leaf names.

Returns a node of the resulting tree, from which it can be traversed.

dm : Matrix used to create Tree.

leaf_names : Array containing names of leaf nodes.

MCPhyloTree.neighbor_joining — Method

neighbor_joining(dm::Array{Float64,2})

This function returns a phylogenetic tree by using neighbor-joining based on a given distance matrix. Creates an array of nodes to be used as leaves.

Returns a node of the resulting tree, from which it can be traversed.

dm : Matrix from which to create tree.

MCPhyloTree.neighbor_joining_int — Method

neighbor_joining_int(dm::Array{Float64,2},leaves::Vector{FNode})

–- INTERNAL –- Internal function that is called by both neighbor_joining methods. Contains the actual neighbor-joining algorithm, i.e. builds a phylogenetic tree from the given distance matrix and array of leaves. Returns a node of that tree, from which it can be traversed.

MCPhyloTree.upgma — Method

upgma(dm::Array{Float64,2}, Array{String,1})

This function returns a phylogenetic tree by using UPGMA based on a given distance matrix and an array of leaf names.

Returns a node of the resulting tree, from which it can be traversed.

dm : Matrix from which to create the tree.

leaf_names : array of strings containing names of leaf nodes.

MCPhyloTree.upgma — Method

upgma(dm::Array{Float64,2})

This function returns a phylogenetic tree by using UPGMA based on a given distance matrix. Creates an array of nodes to be used as leaves.

Returns a node of the resulting tree, from which it can be traversed.

dm : Matrix from which to create the tree.

MCPhyloTree.upgma_int — Method

upgma_int(dm::Array{Float64,2},leaves::Vector{FNode})

–- INTERNAL –- Internal function that is called by both UPGMA methods. Contains the actual UPGMA algorithm, i.e. builds a phylogenetic tree from the given distance matrix and array of leaves. Returns a node of that tree, from which it can be traversed.

Tree Consensus

MCPhyloTree.check_node! — Method

check_node!(
    ref_node::T, leaves::Vector{T},
    leaf_ranks_reverse::Dict{String, Int64},
    xleft_dict::Dict{T, Tuple{T, Dict{Int64, T}}},
    xright_dict::Dict{T, Tuple{T, Dict{Int64, T}}},
    marked_nodes::Union{Dict{Int64, Bool, Nothing}}=nothing
    )::Union{Nothing, Tuple{T, T, T, Dict{Int64, T}, Dict{Int64, T}}} where T<:GeneralNode

–- INTERNAL –- Helper function that handles major chunk of code that onewaycompatible and merge_tree share

MCPhyloTree.count_cluster_occurences — Method

count_cluster_occurences(bit_vectors::BitVector)
    ::PriorityQueue{BitVector, Int64}

–- INTERNAL –- Helper function for greedyconsensustree that counts the occurrences of each bit vector representing a cluster. Returns a priority queue.

MCPhyloTree.depth_dicts — Method

depth_dicts(leaves::Vector{T})
    ::Tuple{Dict{T, Tuple{T, Dict{Int64, T}}}, Dict{T, Tuple{T, Dict{Int64, T}}}} where T<:GeneralNode

–- INTERNAL –- Helper function for onewaycompatible and merge_trees!. Creates a dictionary based on a vector of leaf nodes, and stores the depth of each node, as well as the left and right path leading to it. Based on section 6.1 of the paper.

MCPhyloTree.find_common_clusters — Method

find_common_clusters(ref_tree, tree:T)
    ::Dict{Int64, Tuple{Bool, Union{Float64, Missing}}}

–- INTERNAL –- Use Day's algorithm to create a dictionary, that tells us for each node of the second input tree, if its corresponding cluster is a common cluster of the trees.

Returns dictionary.

ref_tree : Tree used to create dictionary.

tree : Tree used to create dictionary.

MCPhyloTree.get_cluster_start_indeces — Method

get_cluster_start_indeces(ref_nodes::T, tree::T)
    ::Dict{T, Int64} where T<:GeneralNode

–- INTERNAL –- Helper function to obtain the cluster start indeces for a tree (tree), based on the nodes of another tree (ref_nodes).

MCPhyloTree.get_leaf_ranks — Method

function get_leaf_ranks(nodes::Vector{T})
    ::Dict{String, Int64} where T<:GeneralNode

–- INTERNAL –- Enumerate the leaf nodes in a tree. Returns a dictionary of this mapping.

MCPhyloTree.greedy_consensus_tree — Method

greedy_consensus_tree(trees::Vector{T})::T where T<:GeneralNode

Construct the greedy consensus tree from a set of trees that share the same leafset. Returns the root node of the greedy consensus tree, from which it can be traversed. This algorithm is based on section 5 and 6.1 of:

Jesper Jansson, Chuanqi Shen, and Wing-Kin Sung. 2016. Improved algorithms for constructing consensustrees. J. ACM 63, 3, Article 28 (June 2016), 24 pages https://dl.acm.org/doi/pdf/10.1145/2925985

MCPhyloTree.loose_consensus_tree — Method

loose_consensus_tree(trees::Vector{T})::T where T<:GeneralNode

Construct the loose consensus tree from a set of trees that share the same leafset. I.e. a tree with all the clusters that appear in at least one tree and are compatible with all trees. Returns the root node of the loose consensus tree, from which it can be traversed. This algorithm is based on section 4 and 6.1 of:

Jesper Jansson, Chuanqi Shen, and Wing-Kin Sung. 2016. Improved algorithms for constructing consensustrees. J. ACM 63, 3, Article 28 (June 2016), 24 pages https://dl.acm.org/doi/pdf/10.1145/2925985

MCPhyloTree.majority_consensus_tree — Method

majority_consensus_tree(trees::Vector{T}, percentage::Float64=0.5)
    ::T where T<:GeneralNode

Construct the majority rule consensus tree from a set of trees that share the same leafset. By default the output tree includes clusters that occur in over 50% of the trees. This can be customized when calling the function. The function returns the root node of the majority consensus tree, from which it can be traversed. The algorithm is based on section 3 and 6.1 of:

Jesper Jansson, Chuanqi Shen, and Wing-Kin Sung. 2016. Improved algorithms for constructing consensustrees. J. ACM 63, 3, Article 28 (June 2016), 24 pages https://dl.acm.org/doi/pdf/10.1145/2925985

MCPhyloTree.max_leaf_rank — Method

max_leaf_rank(leaf_ranks::Dict{String, Int64}, node::T)
    ::Int64 where T<:GeneralNode

–- INTERNAL –- Recursive helper function to find the highest ranked leaf descendant of a node

MCPhyloTree.merge_trees! — Method

merge_trees!(ref_tree::T, tree::T)::Vector{T}} where T<:GeneralNode

–- INTERNAL –- Merge two compatible trees, i.e. inserts all cluster of the first tree, which aren't already in the second tree, into the second tree

Returns a Tuple containing the root node of the merged tree, as well as a vector of nodes that were added.

ref_tree : tree used to determine common clusters.

tree : tree used to determine common clusters.

MCPhyloTree.min_leaf_rank — Method

min_leaf_rank(leaf_ranks::Dict{String, Int64}, node::T)
    ::Int64 where T<:GeneralNode

–– INTERNAL –- Recursive helper function to find the lowest ranked leaf descendant of a node

MCPhyloTree.one_way_compatible — Method

one_way_compatible(ref_tree::T, tree::T)::T where T<:GeneralNode

–- INTERNAL –- Takes two trees and returns a copy of the first one, where all the clusters that are not compatible with the second tree are removed.

Returns root node of edited tree.

ref_tree : tree used to determine clusters.

tree : tree used to determine clusters.

MCPhyloTree.order_tree! — Method

order_tree!(root::T, cluster_start_indeces::Dict{T, Int64}, leaves=Vector{T}())
    ::Vector{T} where T<:GeneralNode

–- INTERNAL –- Helper function to order a tree based on cluster indeces and return the leaves of the ordered tree

MCPhyloTree.set_node_stats! — Method

set_node_stats!(tree::T, trees::Vector{T}, majority::Bool)
    ::Nothing where T<:GeneralNode

–- INTERNAL –- Helper function for the construction of a consensus tree. Calculates the inc_lengths and statistics of the nodes in the consensus tree. If dealing with a (in-progress) majority consensus tree, this function will also delete its non-majority clusters (handled by the 'majority' boolean).

MCPhyloTree.x_left — Method

x_left(node::T)::Tuple{T,Vector{T}} where T<:GeneralNode

–- INTERNAL –- Helper function to find ancestor of a leaf that has said leaf as leftmost descendant. Also returns the path from the leaf to the mother of that node.

MCPhyloTree.x_right — Method

x_right(node::T)::Tuple{T,Vector{T}} where T<:GeneralNode

–- INTERNAL –- Helper function to find ancestor of a leaf that has said leaf as rightmost descendant. Also returns the path from the leaf to the mother of that node.

Tree Ladderizing

MCPhyloTree.ladderize_tree! — Method

ladderize_tree!(root::T, ascending::Bool=true)
    ::Nothing where T<:GeneralNode

This function ladderizes a tree inplace, i.e. sorts the nodes on all levels by the count of their descendants.

root : root Node of tree.

ascending : Boolean, determines whether to sort in ascending (true) or descending (false) order.

MCPhyloTree.ladderize_tree — Method

ladderize_tree(root::T, ascending::Bool=true)::T where T<:GeneralNode

This function returns a ladderized copy of a tree, i.e. a copy with all the nodes on all levels sorted by the count of their descendants.

root : root Node of tree.

ascending : Boolean, determines whether to sort in ascending (true) or descending (false) order.