Distance

MCPhyloTree.BHV_bounds — Method

BHV_bounds(tree1::T, tree2::T)::Tuple{Float64, Float64} where T <:GeneralNode

This function calculates the lower and upper bounds of the geodesic in the Billera-Holmes-Vogtman space.

Returns tuple of floats.

MCPhyloTree.RF — Method

RF(tree1::T, tree2::T)::Int64 where T <:GeneralNode

Calculate the Robinson-Foulds distance between the two trees. In its current form the function assumes the trees have identical leave sets.

Returns result of algorithm as integer.

tree1 : tree used to determine RF distance.

tree2 : tree used to determine RF distance.

MCPhyloTree.get_bipartitions — Method

get_bipartitions(tree::T)::Vector{Tuple} where T <:GeneralNode

Get a vector of all bipartions of tree.

Returns a vector containing Tuples of sets representing the bipartions.